Cleanup tests
Cleanup and refactor tests
