+++ package.json
@@ -22,8 +22,8 @@
     "generate:meta": "ts-node --skip-project node_modules/.bin/polkadot-types-from-chain --package @interlay/polkabtc/interfaces --endpoint ./src/json/edgeware.json --output ./src/interfaces",
     "test": "run-s build test:*",
     "test:lint": "eslint src --ext .ts",
-    "test:unit": "mocha test/unit/*.test.ts test/unit/**/*.test.ts --timeout 190000",
-    "test:integration": "mocha test/integration/**/*.test.ts",
+    "test:unit": "mocha test/unit/*.test.ts test/unit/**/*.test.ts",
+    "test:integration": "mocha test/integration/**/*.test.ts --timeout 190000",
     "watch:build": "tsc -p tsconfig.json -w",
     "watch:test": "mocha --watch test/**/*.test.ts"
   },
